perf: 合并唱歌与放歌
