+++ .gitconfig
@@ -141,5 +141,3 @@ dmypy.json
 .pytype/
 
 # End of https://www.toptal.com/developers/gitignore/api/python
-
-resource/**
