+++ .gitignore
@@ -140,6 +140,6 @@ dmypy.json
 
 # End of https://www.toptal.com/developers/gitignore/api/python
 
-resource/voices/*
+resource
 accounts
 cache
