--- src/plugins/music/README.md
@@ -1,3 +1,0 @@
-# Music
-
-![考哥！](readme.jpg)
